+++ src/main/java/Lesson_13/PhoneDirectory.java
@@ -1,0 +1,18 @@
+package Lesson_13;
+import java.util.*;
+
+public class PhoneDirectory {
+    private Map<String, List<String>> directory;
+
+    public PhoneDirectory() {
+        this.directory = new HashMap<>();
+    }
+
+    public void add(String surname, String phoneNumber) {
+        directory.computeIfAbsent(surname, k -> new ArrayList<>()).add(phoneNumber);
+    }
+
+    public List<String> get(String surname) {
+        return directory.getOrDefault(surname, Collections.emptyList());
+    }
+}
