+++ src/main/java/Lesson_13/Student.java
@@ -1,0 +1,48 @@
+package Lesson_13;
+
+import java.util.Map;
+
+public class Student {
+    private String name;
+    private String group;
+    private int course;
+    private Map<String, Integer> grades;
+
+    public Student(String name, String group, int course, Map<String, Integer> grades) {
+        this.name = name;
+        this.group = group;
+        this.course = course;
+        this.grades = grades;
+    }
+
+    public double getAverageGrade() {
+        if (grades.isEmpty()) return 0;
+        int sum = 0;
+        for (int grade : grades.values()) {
+            sum += grade;
+        }
+        return (double) sum / grades.size();
+    }
+
+    public void promoteToNextCourse() {
+        this.course++;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getCourse() {
+        return course;
+    }
+
+    @Override
+    public String toString() {
+        return "Student{" +
+                "name='" + name + '\'' +
+                ", group='" + group + '\'' +
+                ", course=" + course +
+                ", grades=" + grades +
+                '}';
+    }
+}
