+++ src/main/java/Lesson_13/StudentProcessor.java
@@ -1,0 +1,22 @@
+package Lesson_13;
+import java.util.Set;
+
+public class StudentProcessor {
+    public static void removeUnderperformingStudents(Set<Student> students) {
+        students.removeIf(student -> student.getAverageGrade() < 3);
+    }
+
+    public static void promoteStudents(Set<Student> students) {
+        for (Student student : students) {
+            if (student.getAverageGrade() >= 3) {
+                student.promoteToNextCourse();
+            }
+        }
+    }
+
+    public static void printStudents(Set<Student> students, int course) {
+        students.stream()
+                .filter(student -> student.getCourse() == course)
+                .forEach(student -> System.out.println(student.getName()));
+    }
+}
